th08: Fix the file_load breakpoint to work with files larger than the original.
In this game, the buffer that holds the final decrypted game data is only allocated in the second, final decompression/decryption function, which makes it impossible to "optimize away" the first round. Interestingly, this function is also called to decompress score.dat and 2 other things, so this breakpoint is hit more often than necessary.

Fortunately though, we clear the file replacement state after every valid breakpoint sequence, so has no negative effect on the game - except for one instance with data dumping where the game would crash when attempting to dump a file that was previously replaced in BP_file_load(). This is fixed in thcrap commit 8a83a5e5e396d8ece4834fe1830924f35a024206.