th08: Don't read file name and file size in the same breakpoint.
Turns out this isn't necessary at all! We're only interested in both reading and writing the file size when the game needs to allocate the buffer, and the file size we would have pulled from the new file_name breakpoint is 4 bytes larger than necessary anyway.

This eliminates the last instance of the '"set_patch_size": false' workaround, and even allows us to remove the file_size_patch breakpoint in the next build, which will further simplify the implementation of file breakpoints.