th14, th143: Fix sprite alignment glitches in the original ascii_960.png.
In order to align all glyphs of the default UI font on the same baseline, the 2nd and 3rd row need to be moved up one pixel, and the 5th row needs to be moved down one pixel.
Also, we move the ^ glyph down one pixel and compress the capital W by one pixel in order to actually fit both inside their sprite boundaries.

Follow-up to https://bitbucket.org/nmlgc/thpatch-bugs/issue/58, thanks to Kill-o-matic for the initial edit.