Import sibgamer's bullet resizing patch.
Can we have 19 more people like DTM? Please?

By all means, if you have a patch and can't be bothered to set up a server to
host it, try contacting anyone in the thcrap patch network. I didn't put those
contact e-mail addresses there for nothing, y'know. Don't count on me reading
MoTK because it's supposed to be "the most important site". I only do that if
I happen to be in the mood for some self-harm.