th11 v1.00a: Actually initialize DirectInput correctly.
First things first: This will break any input issue workarounds that are
*specific* to th11, aren't necessary for any other ZUN game, and
therefore shouldn't have been necessary to begin with.

So, this is what's going on here:
• For some reason, th11 is the only ZUN game that initializes
DirectInput *before* creating its window.
• As part of DirectInput initialization, ZUN calls
IDirect3DInputDevice8::SetCooperativeLevel() on the keyboard device.
• This function requires a window handle. We don't have one yet.
• SetCooperativeLevel() fails, DirectInput initialization aborts as a
result, and the game falls back on the Windows API functions
GetKeyboardState() and joyGetPosEx(), always.

Since GetKeyboardState() is subject to your currently selected system
layout, this now also affects the game's key bindings. For example, if
you happen to use a keyboard layout where the Z key is not next to X,
like QWERTZ or AZERTY, the physical Shot key you have to press in th11
will be different to every other ZUN game.

For the sake of perfect consistency, this fixes the problem the hard way
(as usual…) by replacing the call to the DirectInput init function with
a call to the window creation function, and then tail-calls DirectInput
initialization from there. As a result, SetCooperativeLevel() succeeds,
and you also get the disabled Win key you've come to expect from modern
ZUN games.

In fact, you could have had the effect of applying your system keyboard
layout to the game's key bindings in any other ZUN game too, by checking
the "No DirectInput" option in custom.exe. This option now actually
works for th11 as it does for every other game – previously, it didn't
have any effect due to DirectInput initialization always failing. Which
brings us to the second, forthcoming part of this bugfix.

The th06 version of the string for that option, "入力に DirectInput
を使用しない", is the *only* correct description of what it actually
does. th07 then started adding パッドの in front of that, which has made
that label factually incorrect ever since – it also affects the
keyboard, not just the pad. (Which you would probably never know if you
only ever use QWERTY-based keyboard layouts.) So, we'd better change
that to something more accurate.

However, th07 and th08 place the label in a column which we can't easily
enlarge automatically. This means we'll quickly be running into dialog
space issues if we changed the text to "Use system keyboard layout
instead of DirectInput" or something similar. Therefore, it will be
undoubtedly easier for translators if we just negated both the check box
state and the label as part of thcrap_tsa.dll, and turn the string into
"Use DirectInput (Shot key: %c)". %c would then dynamically be replaced
with the shot key according to the currently selected keyboard layout
whenever the box is unchecked, and with Z whenever it is checked. We
could even leave out the whole "(Shot key: %c)" part if it doesn't
change anyway for the currently selected layout.

Since this string will also be used for people who just want to run the
Japanese games with the rest of the bugfixes provided by base_tsa, we
also need a Japanese translation for it. Splashman gave me "DirectInput
を使用する（ショットキー: %c）"; are there any other Japanese speakers
who can confirm that this is correct and idiomatic?