th07: Added hardcoded strings for results menu
"+Phantasm" strings are presumably used when you unlock phantasm
stage. This probably explains duplicate strings in th08.

Most of those strings are very simillar, but all of them are unique.

As with th08, I have a list of stringdefs for english and japanese,
and will be posting those in a PR.