th09: Correctly center menu item descriptions (except for the very first one).
Yup, the very first menu description you see, the one for Story Mode,
requires a separate hack, which will be an issue for another day. This
happens to be the very first text the game renders, and since this is
pre-th095, the game doesn't call CreateFont() until directly before the
TextOut() call, so we don't have a font to calculate the length with, at
the time the length is supposed to be calculated. Fixing that will
involve faking a text rendering call… somewhere.

I kinda wanted to completely rewrite this whole weirdly interleaved
section of code, but then noticed that there are function calls in
there, which means that any other th09 build would have required a full
new binary hack. So two smaller hacks it is, which can then be easily
backported.