base_tsa: Fix files.js.
Sorry for breaking updates and new installations for the past 68
hours! And since it's going to take a while until we completed the
switch to exclusively using Git for patch updates, I'll now quickly
go write a small verification server using the GitHub Status API so
that this won't happen again.