th09 v1.50a: Re-encrypt dialog strings in-place after showing them.
Thanks to @mistiryshak on Twitter for reporting this. If a dialog
message is shown a second time before being reloaded (looks like this
only happens when game-overing twice on the same stage), decrypting a
XOR'ed string in place actually re-encrypts it again, and previously,
this re-encrypted "garbage" string was what then got rendered.

And since we don't have a string length and the existing dialog string
XOR function ends at the null byte of the *resulting* string, we indeed
have to copy the function and check for the null byte of the *source*
string instead, and hey, look, we can even optimize enough bytes to
inline it there. Hooray.