th09: Get the file size hacks right.
Oh boy, this game. First, there was the completely rewritten Music Room
comment render call code, then there was that one time where I spend 3
days on getting a crash-free instant ending patch and failed, and now
this almost complete file injection rewrite. And we haven't even
touched spell card names yet.

So, what was the problem here? There are two variables that receive the
uncompressed file size (+4 bytes, to include the edz header): [ebx], an
output pointer passed from the caller, and [ebp-4], a local variable of
the file loading function. [ebp-4] is later passed to the LZSS
decompression function, where it is also used to control some
decompression parameters… which means that this *has* to be the original
file size. Any larger, and the last few bytes of the decompressed file
will be corrupted.

So instead of placing a single breakpoint to set our post-JSON size, we
have to place three, at all places where the game actually needs the
size of the final file injected by thcrap. This is why this requires a
new thcrap build with
https://github.com/thpatch/thcrap/commit/21e948c1c2ffd8e093a9cfb421e1c7dccc4db346.

See the TH09 Spanish patch for a test case of where the old behavior
caused a crash.

Some more documentation on the individual hackpoints:

file_size#for_encrypted_buffer
------------------------------
• Placed before the allocation of the buffer that receives the
  *encrypted* original file.
• Needs to be allocated with the post-JSON size from thcrap in case the
  file is fully replaced, in which case this *is* the final buffer.

file_decrypted_buffer_size
--------------------------
• Placed before the allocation of the final buffer that receives the
  *decrypted* original data and serves as input for thcrap's patch
  hooks. We get to this point in the following two cases:
  • Original file + JSON
  • Original file (unpatched)
• The game already subtracted the 4-byte edz header from the file size
  by that point, leaving the true file size in EDI.
• The buffer obviously needs to be allocated with the post-JSON size
  from thcrap.
• Feeding the post-JSON size any earlier (or in fact, writing it into
  EDI or at any earlier point where the value actually ends up in EDI)
  corrupts the LZSS decompression.
→ Therefore, we use a separate register and place a second file_size
  breakpoint (file_size#for_decrypted_buffer) on that same address.