instant_ending: Add support for TH09.
Surprise! Simply starting the game on Stage 9 fixes the white screen
freeze after the credits, for whatever reason. And combined with
dummied-out MSG files (which I first suspected to be the cause of that
freeze), this gives us the "most instant" ending out of all games – you
don't even get to see the in-game interface.

Also, that's one of the most ingenious binhacks in recent memory.

However, for some reason, viewing endings using this patch can not
unlock any pre-Extra Mode characters as you would expect. (That is,
Cirno, Mystia, Lyrica, Tewi, Aya, Medicine, and Yuuka; unlocking Eiki
for Extra as well as the final unlock of all characters for all modes do
work.) Things I've tried to fix this:
• Removing the binhack and simply letting the game through all 9 stages
  with their MSG scripts dummied out like this. Crashes on Stage 6.
• Starting the run at stage 7 or 8 instead.
• Running through stages 1, 5 and 9, by replacing the

	inc [stage]

  instruction with

	add [stage], 4

  This works for unlocking Cirno and Mystia, but won't unlock any
  further characters.