th08: Backport the overhauled file replacement hacks from th09.
Fixes the random th08 crashes introduced with the latest thcrap build.
Turns out that the old hacks did indeed rely on the conditions in
BP_file_size that had to be removed for th09, but only appeared after a
few stages were loaded. Good thing that the .dat file loading code is
equivalent to th09. (Although with some additional ASM bloat due to a
lack of optimization.)