th16: Assume that the ANM format remains unchanged from th11.
If it does, we save a few minutes of me inevitably forgetting to add
this format definition at the right time tomorrow.

If it doesn't, it will blow up for us during sprite boundary dumping
before we ship the file breakpoints, which means that no one will
notice that it's actually the wrong format for this game.