th16 v0.01a: Add safe sprintf (and translation) for high score and replay menus.
We did it! TH16 thcrap support, 100% done on Day 1. Total time:
12 hours, 25 minutes, and 24 seconds. Including breaks and a fair amount
of general fooling around.

While this was the first time we pulled this of, this *was* kind of
slow. ZUN's change to Visual Studio 2015 meant that basically all of the
binary hacks and breakpoints had to be done from scratch.

Notable changes in difficulty as far as thcrap support is concerned:

* Safe sprintf, although it required a new build to be released today,
  has now become as easy as it can possibly be, requiring only two
  instructions to be modified.

* The dialog textbox width calculation now no longer includes absolute
  addresses, which means it's likely to be only one hack for every
  version of a game. (Which is pretty great, as it is one of the more
  complicated hacks.)

* Proper Ruby support got a little bit harder and now requires a custom
  binary hack for each game version.