th095, th125: Enlarge the help sprites to cover the entire playing field.
th095's is right-aligned at 512 pixels, th125's is center-aligned at 320
pixels. Both are scaled down to 77,78% their size, which is why they
have to be 492 pixels wide, rather than 384.

th143's sprites already are large enough by default and even extend a
bit into the HUD. We couldn't do the same in th095 and th125 due to the
scaling factor and the smaller texture width of 512 pixels, and it would
be a bad idea anyway, as the text (that's already pretty small at
640×480) would necessarily have to cover the photos.