th095, th125, th143: Support help message translation.
Turns out that this generic text breakpoint system is a perfect
culmination of everything I've learned about how to best use my own
engine; why didn't I JUST DO IT in 2014 when I conceived it? Sorry for
having kept you waiting for so long.

(Well, even with a dictionary system and no further breakpoints, it
would be possible to retain the ability of translators to always use the
maximum number of possible lines, if we include the number of lines used
by an original message as part of the dictionary, by mapping the first
line to this number, so that we can correctly map the subsequent empty
strings to their translation… but really, this system keeps the number
of breakpoints as low as it can possibly be, and it's not *that* bad.)