th125 v1.00a: Hack around the weird "+11" scaling for the help messages.
Turns out that for all rendered text, TH10-TH13 does something like

	width_of_text_section = max(((width_of_sprite) + 11) * 2, 1024)

before scaling the text section down to `width_of_sprite` pixels for the
sprite texture. Together with the scaling factor of 77.78% on the help
message sprites, this is the reason why the seemingly correct centering
code from the last commit didn't actually center the text exactly.

(See Rx4d32b in th125 v1.00a for the code that does it.)

This +11 addition might seem weird and like a degradation of quality for
no particular reason, but it *does* effectively result in 11 additional
pixels of text, so we shouldn't just outright disable it, because who
knows which translations might actually rely on these pixels. However,
since draw_ctext() is only used for the help messages anyway, it makes
sense to work around it for that particular function.

So, this hack moves the center of the sprite 14 (11 / 0.77778) pixels to
the right, which ends up moving the perceived center of the final scaled
sprite to exactly 320 pixels, where it should be.