th143 v1.00a: Don't show empty day message lines.
Originally, you had to put a backslash as the first character on a day
description line if it shouldn't be rendered and the height of the text
box should be reduced to only fit a single line. But why would a
backslash ever make more sense than a NUL byte? Luckily, the original
text never makes use of this.