th143 v1.00a: Support day description translation.
And with this, TH14.3 is finally supported completely, after three
years.

`"line": 0` is actually important here – without resetting the internal
line counter, the gentext module would try to display lines 3 and 4 when
the player moves from the scene selection back into the day selection.

We're also specifying the `file` again, to make it repatchable in the
same situation.