th10: Locate known hardcoded strings.
Turns out that the TH10 and TH11 versions of the Music Room spoiler 4
and 5 strings actually lack the '。' at the end, and 4 additionally
lacks the '、'. Not really something that would warrant a new string ID.

Also, in the trial version, the `Result Unknown Spell` string uses three
fewer question marks, causing it to appear even more "centered" than the
full version. Since we left-align those strings in all other games, we
also point this one to the regular version.