th09: Don't try to process an ending line after the first non-@ byte.
Ha, the previous hack was immediately broken by the very first Shift-JIS
character in any of the ending scripts: U+3000 IDEOGRAPHIC WHITE SPACE
maps to 0x8140 in Shift-JIS, and 0x40 is the `@` sign in ASCII, which
caused the rest of the line to be interpreted as a script command.

Also needs to be backported to th08 and th07, while th06 probably needs
a different solution due to the Voodoo limit.