instant_ending: th08: Remove all dialogue in Stage 5, Final A and Final B.
Haha, I was such a moron for agreeing to do TH06-TH09 endings before
completing instant_ending support for TH06-TH09, which will require
thecl to have proper de- and recompilation support of the main ECL
functions in these games. Oh well, here's the part that will have to be
done anyway.

Use this together with Skipgame, for now.