th06 v1.02h: Enlarge the width of the text rendering surface to 1024 pixels.
That's it! The "perfect" removal of the "Voodoo limit" I've been chasing
for the last 5 years.

As ingenious as Bruno's solution for the Music Room was (continuing to
render two sprites, but with the same text twice and a negative offset
for the second sprite), it would have only worked for places that
already used two sprites to begin with, namely, Music Room comments and
endings, and not for Music Room titles, in-game music titles, in-game
dialogue, spell cards, stage titles…  which were the entire reason why
this game has been kept in the Sandbox for all that time. Also, it
causes visible artifacts even in the Music Room.

And in the end, it *was* nohing more than a small ZUN bug, where he
pushed the wrong parameter (sprite width rather than text DIB width) to
a function. But hey, it wouldn't have gone smoothly if it wasn't for
those 5 years of additional experience.
Interestingly, it also seems to be the same basic issue that prevents
text in newer Touhou games to be rendered at a higher bit depth, which
means that we now also did an essential bit of work for that feature.

(Now, the *really* perfectionist way would have been to continue using
the sprite width and only apply the "preparation" to the sprite
rectangle by taking the row pitch of the DIB into account… but hey, who
cares about that probably unnoticeable performance drop.)