th06 v0.13: Enlarge the width of the text rendering surface to 1024 pixels.
Now *this* is what I call optimization. Kinda tricky to carve out some
bytes for the bugfix here, but abusing stdcall made it work out nicely.

Pushing the calculated height a second time effectively leaves this
parameter on the stack after the function. This not only saves us a push
for the second call, it also eliminates the need for that local variable
that was previously used to store the calculated height, which leaves
space to push the width of the text rendering surface from the stack
twice.