th06 v1.02h: Correctly scale rotated sprites in 3D space from textures wider than 256 pixels.
Or in other words: ~W I T C H C R A F T~.

What's going on there is that TH06 actually only stores a single quad
from (-128, -128) to (+128, +128) which is moved, scaled, and textured
separately for every single sprite on the screen. In doing that, ZUN
generates both the X/Y scaling factors and the texture coordinates from
the relative texture coordinates calculated from the ANM spec (sprite
width/height divided by texture width/height).
If you now want a larger sprite, 512 pixels wide for example, this no
longer works – since the underlying quad is still 256 pixels wide, the X
value would have to be 2 for the scaling and 1 for the texture
coordinate.

So we have to divide the texture width of every sprite by 256 and
multiply that with the scaling factor *at the time it's calculated*.
This ensures that the texture coordinates remain untouched, since their
original calculation remains valid even with larger sprites. And we're
really lucky that this actually doesn't cause other glitches, because
the only texture that isn't 256 pixels wide in the original game is the
blurred pause menu snapshot framebuffer at 128×128 pixels, which isn't
rendered in 3D space.

And yeah, I'd techically have to do the same thing for the Y coordinate,
but I'd need more space for the ASM code, and we don't need it right
now, unless someone wants to make a high-res patch or something. So it
should probably be a breakpoint instead, but this is pretty much the
hottest innermost loop in the entire game and our breakpoints aren't too
fast yet…

sprite3d_rotated_voodookill
---------------------------
• Keep EDX pointing to the sprite instance. Savings by not reloading
  that pointer three times in this section: 9 bytes.
• After the REP MOVSD above, EDI points directly after the local copy of
  the world matrix. Since we need element [0][0] next, roll EDI back to
  the beginning of the D3DXMATRIX structure by subtracting its size.
  Savings over accessing the field directly through EBP: 5 bytes.
• Mutiply the X scaling value with (texture_width / 256.0f):

  MOV   ECX, [EDX+sprite_instance_t.spec] (6)
  PUSH  256.0f                            (5)
  FLD   [ECX+sprite_spec_t.texture_width] (3)
  FLD   [ESP]                             (3)
  FDIVP ST(1)                             (2)
  FMULP ST(1)                             (2)
  POP   ECX                               (1)

  Cost: 22 bytes.
• Reuse EDI to multiply element [1][1] (at [EDI+14h]) with the sprite's
  current negated Y scaling factor. Savings over accessing the field
  directly through EBP: 6 bytes.
• Replace FCOMP [pointer to a 0.0f constant] with FLDZ+FCOMPP. Savings:
  2 bytes, 1 memory reference.
• The JNP ends up 3 bytes later than in the original binary, so it needs
  to jump 3 bytes fewer.

… and this fixes… only the stage title, while it is animated. Oh well,
just one copy-pasta away from fixing everything else.