th16 v1.00a: Port all existing hacks from the trial.
... except for spell cards, which we did announce would be last anyway.
But this makes all other translated content for the trial show up in the
full verion.

The only real issue was Skipgame, where our existing boss scripts for
the Reitaisai trial crash on the full version.
Which led to the interesting discovery that you can turn Skipgame into a
boss rush by simply ignoring all boss-specific scripts in the games that
have them. :O