th135: replace a GetFullPathName call with a strcpy
With win32_utf8, GetFullPathName returns an utf-8 string.
th135 calls GetFullPathName on a lot of its internal filenames, then
hashs these filenames. Of course, an utf-8 string doesn't have the same
hash than a shift-jis string.
I then normalizes the filename by replacing backslashs with forward
slashes, and seems to remove the current directory from the path if
needed. The result is that, if we remove the GetFullPathName call, we
will have a few layers normalizing the path - and the GetFullPathName
call isn't really needed because these layers does exactly the contrary
of what GetFullPathName does. So replacing with a strcpy works.