th07: Actually allow replacement files to exceed the original file size.
Which now allows PNGs to be used in place of JPEGs in this game as well.
Interesting how this is the only "JPEG game" where the file size hacks
actually *can* be done in a single breakpoint, since the variable we set
here is only used for the right things (the buffer allocation, and the
caller), and actual decryption does a completely new lookup.