th11: Support v1.03 of Riri's practice patch.
Would have been nice to add support for the old versions in case someone
out there is still using them, but they are nowhere to be found.
However, according to the release thread at

	https://tieba.baidu.com/p/4260569009

v1.03 was released within 24 hours of v1.00, so the old one shouldn't
have spread too much.

Crashes with instant_ending when trying to practice the Stage 1 boss,
but I guess that's to be expected. :P