th155: instant_ending: implement in a more stable way.
The old way had a race between the "begin fight" and the "end fight"
events. If the "begin fight" one was first, it worked, but if the "end
fight" one was first, it was followed by the fight initialization, but
the enemy still didn't have any moveset. The game was stuck in a
"between 2 spells" state, and you could move, but you couldn't touch the
enemy.
The new way is more straightforward. Just calling Go_Ending.
I thought of just using a single file, and patching script/scene/story.nut
to load that file every time, but I would have to update the patch for
every version...