th09: Support spell card translation for v1.50a and v0.02a.
So, what's going on there, and why we are only getting support for this
now?
First, the SHT format directly limits us to 64 bytes, which is also the
is way too little if we take UTF-8 into account.

(All absolute addresses below refer to TH09 v1.50a.)

It's fairly clear where the string replacement breakpoints needed to go.
Could have been gentext, but using spell_name instead allows us to
deduplicate some of the identical spell card names for levels 1 and 2.
If we're there, …
• ESI points to an ínstance of the CardAttackInf class
• [ESI+8] points to some structure with player info
• [[ESI+8]+0x20] points to the ID of the player firing the spell card.

Yup, expressions once again. Which we didn't have in thcrap until last
year.

So far, so good, but where do we get the spell *level* from? 0x403E50 is
the "start new card attack" function of the CardAttackInf class, to
which the spell level is passed as a parameter. Now, we could set a
breakpoint at the beginning of the function to track it… except that
this function is called when the attack is *initiated*, which typically
is about a second before the spell title is actually displayed.
That screams race condition.

So, the level is certainly saved somewhere, right? Turns out there are
actually two code paths, both of which track the spell ID in a different
structure field, at +0x1358 and +0x134C respectively.
Sure, IDA says that both fields are only used within this function. But
who knows whether those members aren't part of a smaller embedded
structure that is directly accessed somewhere else, with different,
smaller offsets? Maybe there *was* actually a reason for two separate
fields there. Too much reverse-engineering necessary to disprove that.

The only place we can safely store the spell level is… the string buffer
itself? So let's use the old safe take the first 8 bytes of that and
repurpose them as a pointer to the spell card string in the SHT, and the
spell level number right next to it.

Or actually, let's don't store the level, and just place the spell_name
breakpoint directly where we store the string pointer. Reduces the
binary hack to a single instruction.

Looking at it with more paranoia though, this doesn't *seem* all too
safe either. I mean, sure, this game *thankfully* doesn't copy spells
into score.dat, can we really be sure that the game isn't doing
something else with the spell card name, which would now be broken? But
well, the fact that the card name is completely processed (with no
traces left) so long before it actually is displayed forces us to use
*some* kind of indirection.
And tracking it separately using a new thcrap module would definitely be
overkill.

Touhou patching is a dangerous business.

Funded by DTM.