th06 v1.02h: Derive the width of the dialog box from the first line sprite.
Well then. How perfectionist do we want to be today?

On the one hand, we do want more than 256 pixels of horizontal space for
text, that's the entire reason I've kept support for this game in the
sandbox for the past 6 years, much to the annoyance of translators
everywhere.
On the other hand, it suddenly looks *just weird* having such a long
text box in this game. And shipping that as part of base_tsa, for every
language, including original Japanese…

That basically was the entire conflict that kept me procrastinating this
issue.

But wait. How about *dynamically sizing* the box based on the length of
the dialog line sprite from text.anm? Therefore, the necessary hack
would simply remove the hardcoded assumption of the box being 256 pixels
wide. Later, we could even dynamically adjust that width from thcrap's
MSG patcher, picking the necessary width from the [256; 384] range on a
per-stage basis, to keep the original look where possible.

Best compromise, I'd say. The only small "drawback", if you want to call
it that, is that this always ensures a padding of 16 pixels from the
edge of the box to the text sprite, unlike TH07 where the text can also
"spill" into the padding on the right side of the box.

dialog_box_leftedge
-------------------

• Remove the subtraction with the 256.0f constant. Savings: 6 bytes.
• Remove the needless store and load to that unused variable ([EBP-9C]
  and [EBP-A4], respectively). Savings: 12 bytes.

• Use EAX as a temporary register for storing the pointer to the width
  of the first dialog line, which allows us to easily retrieve this
  value in the _rightedge binary hack below. We retrieve the width from
  the huge sprite manager thingy class that 0x6D4588 points to. This
  class keeps an array of 2048 sprite boundary structures, each 0x38
  bytes large, right at the beginning. The sprite we need is at ID 0x700
  (start offset of the first text.anm sprite) + 2 (sprite # within
  text.anm), the width is at structure offset 0xC.
  So, [EAX+(0x702*0x38)+0xC].

  MOV  EAX, [0x6D4588]    (5)
  LEA  EAX, [EAX+0x1887C] (6)

  Cost: 11 bytes.

• Subtract the float at [EAX].

  FSUB ST(0), [EAX]       (2)

• NOP out the 5 bytes we saved.

`dialog_box_rightedge_1` and `dialog_box_rightedge_2` then simply
replace the hardcoded reference to the 256.0f constant with [EAX].

Funded by DTM.