th06: Enlarge Music Room text sprites to cover the entire screen.
With the existing removal of the Voodoo limit, this is enough to allow
the track names to span the entire screen, just as in later games.

music01.anm describes the sprites for the first 16 track names in the
track selection, music02.anm covers the last 16. Or effectively, just
the 17th, because this game just had to include one more track that
couldn't fit onto a single 256×256 texture where each sprite is 16
pixels high. Yay for duplicated files.

The music comment line sprites are sprites 8-23 in `text.anm`, meaning
that the first 5 lines are actually shared with the ending lines.

Funded by DTM.