th06: Change the Latin script font to Arial.
Boss titles are just unreadable with Aroania in their originally
intended alignment. Translation patches seem to prefer images for this,
but we shouldn't force them, and Arial improves readability quite a bit.
The static English patch was right all along, as far as font choices are
concerned…

Good thing that everybody listened to me, nobody has *seriously
translated* this game yet, and we can just change to a font that'd give
us significantly less space on a whim… *right*?!

Funded by DTM.