th07 v1.00b: Support optional images for boss titles in place of plaintext.
The entire challenge here being the fact that we want three nameplates
for the three Prismrivers in Stage 4. Ultimately, this giant expression
translates to

	stage_number - 1 + ((stage_number >= 5) * 2) + (
		(stage_number == 4) * (
			huge_class_containing_the_msg_parser->
			boss_portrait_sprite_instance.sprite_id - (
				BOSS_FACE_BASE_SPRITE_ID / 3
	)	)	)

and abuses that
1) the original MSG scripts always set the boss portraits using opcode 2
   before showing the title lines with opcode 8, and
2) each sister has three portraits, which come in the defined
   Lunasa→Merlin→Lyrica order.

We only apply this face lookup if we are in Stage 4, since it will break
in the Extra Stage for Ran otherwise.

Encoding some hidden layout markup in the .msg file to precisely
communicate the character might have resulted in a shorter expression,
but at the cost of complexity in both the patch files and the MSG
patcher, since we need a language-independent way of encoding those IDs.
(Remember that the order they are shown in depends on the player
character.) And suddenly, referring to the portrait sprites to derive
the nameplate does seem like The Absolutely Right Thing To Do.

Check the TH07 image page at Touhou Patch Center for more info on the
format of these images:

	https://www.thpatch.net/wiki/Th07/Images#.28Custom.29_Character_titles

Funded by Splashman.