th13: Add a breakpoint for capturing the trance BGM seek offset.
Necessary for the upcoming support for BGM modding.

To cover all games with a minimum of code, we theoretically only need to
hook SetFilePointer() to catch all BGM switches for ≥TH07. Then, we
simply compare the seek offset to the track start positions from
thbgm.fmt, and have our filenames.

But with the BGM switches that occur when entering and leaving trance
mode in TH13, SetFilePointer() only ever sees the sum of the track start
position and the offset into the track. Since we obviously want to allow
modded BGM to be longer (and larger) than the original, this already
would necessitate mapping the start positions of modded BGM in the
thbgm.fmt structure to the "empty space after the end of thbgm.dat",
thereby imposing a theoretical limit on BGM length. Sure, not too nice,
but also not to big of a deal in practice.

The real problem occurs when non-trance and trance tracks have different
lengths. In that case, the aforementioned sum still ends up pointing
into an unintended, different track. What should we do then? Calculate
the maximum track length and make sure that all tracks occupy that
virtual amount of space after the end of thbgm.dat? Then add further
game-specific knowledge to only build the maximum for each non-trance /
trance track pair? Sure, but… eh, that's limits on top of limits.

And sure, while breakpoints always seem like a maintenance burden, this
is way cleaner than any alternative I could come up with. Requires no
`game_id == TH13` branches in the C++ code, we simply capture the
track's offset, subtract it from SetFilePointer()'s distance parameter
to identify the track, and can still seek the track to the position the
game wants. Done, and without limits.

Funded by Splashman.