th13: Also capture the correct seek offset after the track looped once.
Turns out there actually is a second branch, where the loop size is
subtracted from the seek offset passed to the BGM switch function that
we previously captured. Not capturing this adjusted seek offset might
seek us "outside" of a modded track in that case…

Anyway, can we be done with BGM modding now? Please?

Funded by Splashman.