th095: Make sure that help images are fully patchable.
Tried an algorithmic solution to the problem of patching those THTX regions not
covered by sprites defined in the ANM header, but it turns out that due to
certain junk sprites in ≥TH12's front00.png, this would cause just as many
problems as it would have solved, if not more. Oh, and that algorithm would
have also run in Ο(n²) with the number of sprites. Suddenly, handling this
manually for those few instances where it's a problem doesn't seem so bad
anymore…
Thanks to Redhawk for reporting this issue.

Funded by rosenrose.