th06: Exchange fixed Spell Card results for slightly glitched Music Room text.
Turns out that the odd-numbered sprites actually are used in the former,
and we can't just give them a width of 0 to avoid them being used in the
Music Room.
So, the Voodoo limit removal in the latter would need to be completely
redone to only use one sprite per line. But apparently, it's not such a
big deal among the people I asked, so…