th06: Enforce UTF-8 for the Japanese in-game BGM format and BGM titles.
Which 1) allows patches that don't want to use images to translate
either one and not end up with mojibake, and 2) still doesn't cause
mojibake when running just base_tsa without a translation, as it now
contains both parts of the combined string.

The drawback of this rather dumb workaround is, of course, that we undo
BGM translations made by underlying static patches. What we *actually*
would have wanted there would be automatic conversion from the fallback
encoding to UTF-8 in thcrap's strings_sprintf().
Oh, and Japanese→target language dictionaries… *but*:
• I wasn't funded for that much, and really need to stop blowing these
  pushes out of proportion. Heck, 42d870e already took way longer than
  the cheap bonus it was supposed to be.
• I've looked at the code for a while, and concluded that I'd basically
  have to do both the conversion *and* the dictionary lookup at the same
  time if I don't want to rewrite it all later.
• Especially since strings_sprintf() is pretty hot code and the result
  should at least run decently fast.
• Oh, and the static Russian patch exists, reminding me once again that
  stringlocs weren't the terrible idea they often seem to be.

Funded by zorg.