th06, th07, th08: Force the HUD to be redrawn every frame.
You actually do *not* want to flip that flag bit in the .cfg structure,
because in TH06, that also affects the rest of the backbuffer, slightly
changing the fog color. Q17 in マニュアル/html/faq.html suggests that
not clearing produces the actually intended effect, so we really don't
want to mess with the lore there.
That leaves TH07 and TH08, where the corresponding custom.exe option
actually refers to the HUD exclusively. Which means that we could now
build an elaborate flag flipping system into thcrap_tsa. modeled after
C's #pragma stacks, combined with hooks in the TL note system… or we
could just force-redraw every frame, and no one would probably notice.
…yup, that's what we do, TL notes have been taking way too long
already.

Funded by zorg.