th10: replay and scorefile name stringlocs
Though I doubt the usefulness of "th10_replay_user_find" because th10 doesn't have a page for displaying user replays in it's replay menu. However the existence of that string and the call to FindFirstFile -> FindNextFile do suggest that ZUN did want to have that functionallity.
This could mean that most of the code required to display a user replay page is already there and it wouldn't take that much code to add a user replay page to the th10 replay menu