Add th155 v1.10_steam
The Steam version of Touhou 15.5 have an up to date th155b.pak, but it has the
v1.02b th155.exe (this works because almost all the game engine is in th155.pak
and th155b.pak). This meant that this version was patched as if it was the
v1.02b, which means the players had the "game is outdated" warning at startup,
and, more importantly, our v1.02b support was quite far behind in terms of
features...

So I removed the v1.02b version (I don't think anyone is still using it anyway),
and I added a v1.10_steam, with:
- A base_tasofro/th155.v1.10_steam.js based on the v1.02b one, because this is
the v1.02b exe and the binhacks are still correct.
- All the files in */th155/* copied from the v1.10 version, because these apply
to th155b.dat and this one comes from v1.10.
- A new script_latin/th155.v1.10_steam.js with the correct binhacks and
breakpoints for the v1.02 exe, because we didn't have them (I worked on the
th155 fonts when v1.02b was already outdated).